@@ -7,7 +7,7 @@ version: 0.7.0
 homepage: https://freshchat.com
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=1.10.0"
 
 dependencies:
